isl_multi_templ.c: extract out isl_multi_zero_templ.c
This makes the function isl_multi_*_zero opt-in rather than opt-out.
