+++ isl_multi_templ.c
@@ -604,43 +604,6 @@ error:
 	return NULL;
 }
 
-#ifndef NO_ZERO
-/* Construct a multi expression in the given space with value zero in
- * each of the output dimensions.
- */
-__isl_give MULTI(BASE) *FN(MULTI(BASE),zero)(__isl_take isl_space *space)
-{
-	int n;
-	MULTI(BASE) *multi;
-
-	if (!space)
-		return NULL;
-
-	n = isl_space_dim(space , isl_dim_out);
-	multi = FN(MULTI(BASE),alloc)(isl_space_copy(space));
-
-	if (!n)
-		isl_space_free(space);
-	else {
-		int i;
-		isl_local_space *ls;
-		EL *el;
-
-		space = isl_space_domain(space);
-		ls = isl_local_space_from_space(space);
-		el = FN(EL,zero_on_domain)(ls);
-
-		for (i = 0; i < n; ++i)
-			multi = FN(FN(MULTI(BASE),set),BASE)(multi, i,
-							    FN(EL,copy)(el));
-
-		FN(EL,free)(el);
-	}
-
-	return multi;
-}
-#endif
-
 __isl_give MULTI(BASE) *FN(MULTI(BASE),drop_dims)(
 	__isl_take MULTI(BASE) *multi,
 	enum isl_dim_type type, unsigned first, unsigned n)
